// testapps/nextjs-relay/__generated__/pages_indexQuery.graphql.ts
export type pages_indexQuery$variables = {
	city: string;
};
export type pages_indexQuery$data = {
	readonly weather_getCityByName: {
		readonly weather: {
			readonly summary: {
				readonly ' $fragmentSpreads': FragmentRefs<'Weather_Details'>;
			} | null;
			readonly temperature: {
				readonly ' $fragmentSpreads': FragmentRefs<'Temperature_Details'>;
			} | null;
		} | null;
	} | null;
};
export type pages_indexQuery = {
	response: pages_indexQuery$data;
	variables: pages_indexQuery$variables;
};